Rename scooter hub types module to avoid stdlib clash
