+++ tools/scooter_sensor_hub/models.py
